// frontend/src/types/auth.types.ts
export interface LoginFormInputs {
    email: string;
    password: string;
};

export interface RegisterFormInputs {
    username: string;
    email: string;
    password: string;
    confirmPassword: string;
};

export interface AccountCredentials {
    username: string;
    email: string;
    password: string;
};